+++ integration_test/types/assets.d.ts
@@ -1,4 +1,0 @@
-declare module '*.vtt' {
-  const assetId: number;
-  export default assetId;
-}